# trojan,pss: optimise trojan mining