--- a/pkg/pss/pss_test.go
+++ b/pkg/pss/pss_test.go
@@ -123,7 +123,7 @@ func TestDeliver(t *testing.T) {
 	// test chunk
 	target := trojan.Target([]byte{1}) // arbitrary test target
 	targets := trojan.Targets([]trojan.Target{target})
-	c, err := msg.Wrap(targets)
+	c, err := msg.Wrap(ctx, targets)
 	if err != nil {
 		t.Fatal(err)
 	}