--- a/pkg/trojan/error.go
+++ b/pkg/trojan/error.go
@@ -15,9 +15,6 @@ var (
 	// ErrVarLenTargets is returned when the given target list for a trojan chunk has addresses of different lengths
 	ErrVarLenTargets = errors.New("target list cannot have targets of different length")
 
-	// ErrUnMarshallingTrojanMessage is returned when a trojan message could not be de-serialized
+	// ErrUnmarshal is returned when a trojan message could not be de-serialized
 	ErrUnmarshal = errors.New("trojan message unmarshall error")
-
-	// ErrMinerTimeout is returned when mining a new nonce takes more time than swarm.TrojanMinerTimeout seconds
-	ErrMinerTimeout = errors.New("miner timeout error")
 )