--- a/pkg/trojan/export_test.go
+++ b/pkg/trojan/export_test.go
@@ -1,7 +1,5 @@
 package trojan
 
 var (
-	Contains  = contains
-	HashBytes = hashBytes
-	PadBytes  = padBytesLeft
+	Contains = contains
 )