closes #444 

 - introduce context to Wrap and eliminate MinerTimeout
 - simplify Wrap
 - add mine function that finds a nonce that makes a function true
 - introduce mining benchmarks
BEFORE
```
:) go test -v -run None -bench .
goos: darwin
goarch: amd64
pkg: github.com/ethersphere/bee/pkg/trojan
BenchmarkWrap
BenchmarkWrap/length:1,depth:1
BenchmarkWrap/length:1,depth:1-8         	     100	  25259359 ns/op
BenchmarkWrap/length:4,depth:1
BenchmarkWrap/length:4,depth:1-8         	     211	   5987303 ns/op
BenchmarkWrap/length:16,depth:1
BenchmarkWrap/length:16,depth:1-8        	     662	   1601796 ns/op
BenchmarkWrap/length:64,depth:2
BenchmarkWrap/length:64,depth:2-8        	      15	 118159956 ns/op
BenchmarkWrap/length:256,depth:2
BenchmarkWrap/length:256,depth:2-8       	      44	  23306940 ns/op
BenchmarkWrap/length:1024,depth:2
BenchmarkWrap/length:1024,depth:2-8      	     188	   6117949 ns/op
BenchmarkWrap/length:1024,depth:3
BenchmarkWrap/length:1024,depth:3-8      	       1	4418328590 ns/op
BenchmarkWrap/length:4096,depth:3
BenchmarkWrap/length:4096,depth:3-8      	       3	 443126126 ns/op
BenchmarkWrap/length:16384,depth:3
BenchmarkWrap/length:16384,depth:3-8     	       6	 203879796 ns/op
PASS
ok  	github.com/ethersphere/bee/pkg/trojan	21.083s
```

AFTER
```
goarch: amd64
pkg: github.com/ethersphere/bee/pkg/trojan
BenchmarkWrap
BenchmarkWrap/length:1,depth:1
BenchmarkWrap/length:1,depth:1-8         	      93	  13613038 ns/op
BenchmarkWrap/length:4,depth:1
BenchmarkWrap/length:4,depth:1-8         	     325	   4058203 ns/op
BenchmarkWrap/length:16,depth:1
BenchmarkWrap/length:16,depth:1-8        	     705	   1583214 ns/op
BenchmarkWrap/length:16,depth:2
BenchmarkWrap/length:16,depth:2-8        	       8	 137261977 ns/op
BenchmarkWrap/length:64,depth:2
BenchmarkWrap/length:64,depth:2-8        	      27	  49296687 ns/op
BenchmarkWrap/length:256,depth:2
BenchmarkWrap/length:256,depth:2-8       	      90	  13566495 ns/op
BenchmarkWrap/length:256,depth:3
BenchmarkWrap/length:256,depth:3-8       	       1	1618499838 ns/op
BenchmarkWrap/length:4096,depth:3
BenchmarkWrap/length:4096,depth:3-8      	       8	 265504439 ns/op
BenchmarkWrap/length:16384,depth:3
BenchmarkWrap/length:16384,depth:3-8     	      26	  78521115 ns/op
PASS
ok  	github.com/ethersphere/bee/pkg/trojan	16.418s

```

at https://github.com/ethersphere/bee/pull/695/commits/12b4f9779515aa7aa9789b0bb9d61ec208d905ca

```
goos: darwin
goarch: amd64
pkg: github.com/ethersphere/bee/pkg/trojan
BenchmarkWrap
BenchmarkWrap/length:1,depth:1
BenchmarkWrap/length:1,depth:1-8         	     100	  13093521 ns/op
BenchmarkWrap/length:4,depth:1
BenchmarkWrap/length:4,depth:1-8         	     330	   4050533 ns/op
BenchmarkWrap/length:16,depth:1
BenchmarkWrap/length:16,depth:1-8        	     722	   1580087 ns/op
BenchmarkWrap/length:16,depth:2
BenchmarkWrap/length:16,depth:2-8        	      40	 236741716 ns/op
BenchmarkWrap/length:64,depth:2
BenchmarkWrap/length:64,depth:2-8        	      26	  49744130 ns/op
BenchmarkWrap/length:256,depth:2
BenchmarkWrap/length:256,depth:2-8       	     100	  16560863 ns/op
BenchmarkWrap/length:256,depth:3
BenchmarkWrap/length:256,depth:3-8       	       2	1991208560 ns/op
BenchmarkWrap/length:4096,depth:3
BenchmarkWrap/length:4096,depth:3-8      	       3	 400150111 ns/op
BenchmarkWrap/length:16384,depth:3
BenchmarkWrap/length:16384,depth:3-8     	      13	  93524278 ns/op
```